Separate registration routes from main back-end class
