+++ bot/src/main/kotlin/org/epilink/bot/http/LinkBackEnd.kt
@@ -21,7 +21,6 @@ import io.ktor.http.HttpStatusCode
 import io.ktor.http.ParametersBuilder
 import io.ktor.http.content.TextContent
 import io.ktor.jackson.jackson
-import io.ktor.request.ContentTransformationException
 import io.ktor.request.header
 import io.ktor.request.receive
 import io.ktor.response.respond
@@ -31,12 +30,13 @@ import kotlinx.coroutines.coroutineScope
 import org.epilink.bot.*
 import org.epilink.bot.StandardErrorCodes.*
 import org.epilink.bot.config.LinkWebServerConfiguration
-import org.epilink.bot.db.Disallowed
 import org.epilink.bot.db.LinkServerDatabase
 import org.epilink.bot.db.LinkUser
 import org.epilink.bot.db.UsesTrueIdentity
 import org.epilink.bot.discord.LinkRoleManager
-import org.epilink.bot.http.data.*
+import org.epilink.bot.http.data.InstanceInformation
+import org.epilink.bot.http.data.RegistrationAuthCode
+import org.epilink.bot.http.data.UserInformation
 import org.epilink.bot.http.sessions.ConnectedSession
 import org.epilink.bot.http.sessions.RegisterSession
 import org.epilink.bot.ratelimiting.RateLimiting
@@ -54,6 +54,8 @@ interface LinkBackEnd {
      * Ktor module for the back-end API
      */
     fun Application.epilinkApiModule()
+
+    fun ApplicationCall.loginAs(user: LinkUser, username: String, avatar: String?)
 }
 
 /**
@@ -81,6 +83,8 @@ internal class LinkBackEndImpl : LinkBackEnd, KoinComponent {
     private val legal: LinkLegalTexts by inject()
 
     private val wsCfg: LinkWebServerConfiguration by inject()
+
+    private val registrationApi: LinkRegistrationApi by inject()
 
     override fun Application.epilinkApiModule() {
         /*
@@ -110,6 +114,7 @@ internal class LinkBackEndImpl : LinkBackEnd, KoinComponent {
         routing {
             route("/api/v1") {
                 epilinkApiV1()
+                registrationApi.install(this)
             }
         }
     }
@@ -238,89 +243,6 @@ internal class LinkBackEndImpl : LinkBackEnd, KoinComponent {
                 }
             }
         }
-
-        route("register") {
-            rateLimited(limit = 10, timeBeforeReset = Duration.ofMinutes(1)) {
-                @ApiEndpoint("GET /api/v1/register/info")
-                get("info") {
-                    val session = call.sessions.getOrSet { RegisterSession() }
-                    call.respondRegistrationStatus(session)
-                }
-
-                @ApiEndpoint("DELETE /api/v1/register")
-                delete {
-                    logger.debug { "Clearing registry session ${call.request.header("RegistrationSessionId")}" }
-                    call.sessions.clear<RegisterSession>()
-                    call.respond(HttpStatusCode.OK)
-                }
-
-                @ApiEndpoint("POST /api/v1/register")
-                post {
-                    with(call.sessions.get<RegisterSession>()) {
-                        if (this == null) {
-                            logger.debug {
-                                "Missing/unknown session header for call from reg. session ${call.request.header("RegistrationSessionId")}"
-                            }
-                            call.respond(
-                                HttpStatusCode.BadRequest,
-                                ApiErrorResponse("Missing session header", IncompleteRegistrationRequest.toErrorData())
-                            )
-                        } else if (discordId == null || discordUsername == null || email == null || microsoftUid == null) {
-                            logger.debug {
-                                """
-                            Incomplete registration process for session ${call.request.header("RegistrationSessionId")}
-                            discordId = $discordId
-                            discordUsername = $discordUsername
-                            email = $email
-                            microsoftUid = $microsoftUid
-                            """.trimIndent()
-                            }
-                            call.respond(
-                                HttpStatusCode.BadRequest,
-                                ApiErrorResponse(
-                                    "Incomplete registration process",
-                                    IncompleteRegistrationRequest.toErrorData()
-                                )
-                            )
-                        } else {
-                            val regSessionId = call.request.header("RegistrationSessionId")
-                            logger.debug { "Completing registration session for $regSessionId" }
-                            val options: AdditionalRegistrationOptions =
-                                call.receiveCatching() ?: return@post
-                            val u = db.createUser(discordId, microsoftUid, email, options.keepIdentity)
-                            roleManager.invalidateAllRoles(u.discordId)
-                            call.loginAs(u, discordUsername, discordAvatarUrl)
-                            logger.debug { "Completed registration session. $regSessionId logged in and reg session cleared." }
-                            call.respond(HttpStatusCode.Created, apiSuccess("Account created, logged in."))
-                        }
-                    }
-                }
-
-                @ApiEndpoint("POST /register/authcode/discord")
-                @ApiEndpoint("POST /register/authcode/msft")
-                post("authcode/{service}") {
-                    when (val service = call.parameters["service"]) {
-                        null -> error("Invalid service") // Should not happen
-                        "discord" -> processDiscordAuthCode(call, call.receive())
-                        "msft" -> processMicrosoftAuthCode(call, call.receive())
-                        else -> {
-                            logger.debug { "Attempted to register under unknown service $service" }
-                            call.respond(
-                                HttpStatusCode.NotFound,
-                                ApiErrorResponse("Invalid service: $service", UnknownService.toErrorData())
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    private suspend inline fun <reified T : Any> ApplicationCall.receiveCatching(): T? = try {
-        receive()
-    } catch (ex: ContentTransformationException) {
-        logger.error("Incorrect input from user", ex)
-        null
     }
 
     /**
@@ -337,95 +259,16 @@ internal class LinkBackEndImpl : LinkBackEnd, KoinComponent {
             contacts = wsCfg.contacts
         )
 
-    /**
-     * Take a Discord authorization code, consume it and apply the information retrieve form it to the current
-     * registration session
-     */
-    private suspend fun processDiscordAuthCode(call: ApplicationCall, authcode: RegistrationAuthCode) {
-        val session = call.sessions.getOrSet { RegisterSession() }
-        // Get token
-        logger.debug { "Get Discord token from authcode ${authcode.code}" }
-        val token = discordBackEnd.getDiscordToken(authcode.code, authcode.redirectUri)
-        // Get information
-        val (id, username, avatarUrl) = discordBackEnd.getDiscordInfo(token)
-        logger.debug { "Processing Discord info for registration for $id ($username)" }
-        val user = db.getUser(id)
-        if (user != null) {
-            logger.debug { "User already exists: logging in" }
-            call.loginAs(user, username, avatarUrl)
-            call.respond(ApiSuccessResponse("Logged in", RegistrationContinuation("login", null)))
-        } else {
-            val adv = db.isDiscordUserAllowedToCreateAccount(id)
-            if (adv is Disallowed) {
-                logger.debug { "Discord user $id cannot create account: " + adv.reason }
-                call.respond(
-                    HttpStatusCode.BadRequest,
-                    ApiErrorResponse(adv.reason, AccountCreationNotAllowed.toErrorData())
-                )
-                return
-            }
-            logger.debug { "Discord registration information OK: continue " }
-            val newSession = session.copy(discordUsername = username, discordId = id, discordAvatarUrl = avatarUrl)
-            call.sessions.set(newSession)
-            call.respond(
-                ApiSuccessResponse(
-                    "Connected to Discord",
-                    RegistrationContinuation("continue", newSession.toRegistrationInformation())
-                )
-            )
-        }
-    }
 
-    /**
-     * Take a Microsoft authorization code, consume it and apply the information retrieve form it to the current
-     * registration session
-     */
-    private suspend fun processMicrosoftAuthCode(call: ApplicationCall, authcode: RegistrationAuthCode) {
-        val session = call.sessions.getOrSet { RegisterSession() }
-        logger.debug { "Get Microsoft token from authcode ${authcode.code}" }
-        // Get token
-        val token = microsoftBackEnd.getMicrosoftToken(authcode.code, authcode.redirectUri)
-        // Get information
-        val (id, email) = microsoftBackEnd.getMicrosoftInfo(token)
-        logger.debug { "Processing Microsoft info for registration for $id ($email)" }
-        val adv = db.isMicrosoftUserAllowedToCreateAccount(id, email)
-        if (adv is Disallowed) {
-            logger.debug { "Microsoft user $id ($email) is not allowed to create an account: " + adv.reason }
-            call.respond(
-                HttpStatusCode.BadRequest,
-                ApiErrorResponse(adv.reason, AccountCreationNotAllowed.toErrorData())
-            )
-            return
-        }
-        logger.debug { "Microsoft registration information OK: continue " }
-        val newSession = session.copy(email = email, microsoftUid = id)
-        call.sessions.set(newSession)
-        call.respond(
-            ApiSuccessResponse(
-                "Connected to Microsoft",
-                RegistrationContinuation("continue", newSession.toRegistrationInformation())
-            )
-        )
-    }
 
     /**
      * Setup the sessions to log in the passed user object
      */
-    private fun ApplicationCall.loginAs(user: LinkUser, username: String, avatar: String?) {
+    override fun ApplicationCall.loginAs(user: LinkUser, username: String, avatar: String?) {
         logger.debug { "Logging ${user.discordId} ($username) in" }
         sessions.clear<RegisterSession>()
         sessions.set(ConnectedSession(user.discordId, username, avatar))
     }
-
-    private suspend fun ApplicationCall.respondRegistrationStatus(session: RegisterSession) {
-        respond(ApiSuccessResponse(null, session.toRegistrationInformation()))
-    }
-
-    /**
-     * Utility function for transforming a session into the JSON object that is returned by the back-end's APIs
-     */
-    private fun RegisterSession.toRegistrationInformation() =
-        RegistrationInformation(email = email, discordAvatarUrl = discordAvatarUrl, discordUsername = discordUsername)
 
     @UsesTrueIdentity
     private suspend fun ConnectedSession.toUserInformation() =
